updated the APIs for u_digit and u_forDigit to be consistent.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/611



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@6802 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/i18n/rbt_pars.cpp
@@ -1451,7 +1451,7 @@ int32_t TransliteratorParser::parseInteger(const UnicodeString& rule, int32_t& p
     }
 
     while (p < limit) {
-        int8_t d = u_digit(rule.charAt(p++), radix);
+        int32_t d = u_digit(rule.charAt(p++), radix);
         if (d < 0) {
             --p;
             break;
